requirements.txt
streamlit
plotly
pandas
python-dotenv
requests
base64
pathlib
typing
matplotlib
numpy
Pillow
opencv-python-headless
scikit-learn
scipy
seaborn
protobuf
streamlit-extras
st-pages